+++ app/Models/Question.php
@@ -15,8 +15,8 @@ use App\Parents\Model;
  * @property string|null $text
  * @property int $type
  * @property int $files_number
- * @property int $min_points
- * @property int $max_points
+ * @property float $min_points
+ * @property float $max_points
  * @property \App\Models\Group $group
  * @property \Illuminate\Support\Collection|\App\Models\File[] $files
  * @property \Ramsey\Collection\Collection|\App\Models\Option[] $options
@@ -59,8 +59,8 @@ final class Question extends Model implements HasFileTraitInterface
         self::ATTR_TEXT => 'string',
         self::ATTR_TYPE => 'int',
         self::ATTR_FILES_NUMBER => 'int',
-        self::ATTR_MIN_POINTS => 'int',
-        self::ATTR_MAX_POINTS => 'int',
+        self::ATTR_MIN_POINTS => 'float',
+        self::ATTR_MAX_POINTS => 'float',
     ];
 
     /**
@@ -94,15 +94,15 @@ final class Question extends Model implements HasFileTraitInterface
     /**
      * Get real value of min points, considering points for options
      *
-     * @param int $value
-     * @return int
+     * @param float $value
+     * @return float
      */
-    public function getMinPointsAttribute(int $value): int
+    public function getMinPointsAttribute(float $value): float
     {
         if ($this->type === QuestionTypesEnum::OPTIONS_CHECKBOX) {
             return (int)\array_sum(\array_merge([0], $this->options->filter(static function (Option $option): bool {
                 return $option->points < 0;
-            })->map(static function (Option $option): int {
+            })->map(static function (Option $option): float {
                 return $option->points;
             })->toArray()));
         }
@@ -110,7 +110,7 @@ final class Question extends Model implements HasFileTraitInterface
         if ($this->type === QuestionTypesEnum::OPTIONS) {
             return (int)\min(\array_merge([0], $this->options->filter(static function (Option $option): bool {
                 return $option->points < 0;
-            })->map(static function (Option $option): int {
+            })->map(static function (Option $option): float {
                 return $option->points;
             })->toArray()));
         }
@@ -121,15 +121,15 @@ final class Question extends Model implements HasFileTraitInterface
     /**
      * Get real value of max points, considering points for options
      *
-     * @param int $value
-     * @return int
+     * @param float $value
+     * @return float
      */
-    public function getMaxPointsAttribute(int $value): int
+    public function getMaxPointsAttribute(float $value): float
     {
         if ($this->type === QuestionTypesEnum::OPTIONS_CHECKBOX) {
             return (int)\array_sum(\array_merge([0], $this->options->filter(static function (Option $option): bool {
                 return $option->points > 0;
-            })->map(static function (Option $option): int {
+            })->map(static function (Option $option): float {
                 return $option->points;
             })->toArray()));
         }
@@ -137,7 +137,7 @@ final class Question extends Model implements HasFileTraitInterface
         if ($this->type === QuestionTypesEnum::OPTIONS) {
             return (int)\max(\array_merge([0], $this->options->filter(static function (Option $option): bool {
                 return $option->points > 0;
-            })->map(static function (Option $option): int {
+            })->map(static function (Option $option): float {
                 return $option->points;
             })->toArray()));
         }
